# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type MedicineStatus = typeof MedicineStatus[keyof typeof MedicineStatus];


export const MedicineStatus = {
  safe: 'safe',
  near_expiry: 'near_expiry',
  expiring_soon: 'expiring_soon',
  expired: 'expired',
} as const;

export interface MedicineInput {
  name: string;
  batchNumber: string;
  manufacturer: string;
  manufacturingDate: string;
  expiryDate: string;
  quantity: number;
  storageLocation: string;
  category: string;
  qrCodeValue: string;
}

export type Medicine = MedicineInput & {
  id: string;
  remainingDays: number;
  status: MedicineStatus;
  createdAt: string;
};

export interface DashboardSummary {
  total: number;
  safe: number;
  nearExpiry: number;
  expiringSoon: number;
  expired: number;
}

export type ReportsSummaryMonthlySummaryItem = {
  month: string;
  count: number;
};

export interface ReportsSummary {
  expiringThisMonth: number;
  expired: number;
  total: number;
  safe: number;
  nearExpiry: number;
  totalQuantity: number;
  monthlySummary: ReportsSummaryMonthlySummaryItem[];
}

export type ListMedicinesParams = {
search?: string;
status?: MedicineStatus;
sortBy?: ListMedicinesSortBy;
};

export type ListMedicinesSortBy = typeof ListMedicinesSortBy[keyof typeof ListMedicinesSortBy];


export const ListMedicinesSortBy = {
  name: 'name',
  expiryDate: 'expiryDate',
  manufacturer: 'manufacturer',
  quantity: 'quantity',
  createdAt: 'createdAt',
} as const;

export type GetRecentMedicinesParams = {
limit?: number;
};