# lib/api-zod/src/generated/types/medicine.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MedicineInput } from './medicineInput';
import type { MedicineStatus } from './medicineStatus';

export type Medicine = MedicineInput & {
  id: string;
  remainingDays: number;
  status: MedicineStatus;
  createdAt: Date;
};