# lib/api-zod/src/generated/types/medicineInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MedicineInput {
  name: string;
  batchNumber: string;
  manufacturer: string;
  manufacturingDate: Date;
  expiryDate: Date;
  quantity: number;
  storageLocation: string;
  category: string;
  qrCodeValue: string;
}