# lib/api-zod/src/generated/types/medicineStatus.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type MedicineStatus = typeof MedicineStatus[keyof typeof MedicineStatus];


export const MedicineStatus = {
  safe: 'safe',
  near_expiry: 'near_expiry',
  expiring_soon: 'expiring_soon',
  expired: 'expired',
} as const;