# lib/api-zod/src/generated/types/reportsSummary.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReportsSummaryMonthlySummaryItem } from './reportsSummaryMonthlySummaryItem';

export interface ReportsSummary {
  expiringThisMonth: number;
  expired: number;
  total: number;
  safe: number;
  nearExpiry: number;
  totalQuantity: number;
  monthlySummary: ReportsSummaryMonthlySummaryItem[];
}